[Tree] Introduce tree repository traits and remove abstract tree repositories
