+++ composer.json
@@ -38,7 +38,7 @@
         "wiki": "https://github.com/Atlantic18/DoctrineExtensions/tree/master/doc"
     },
     "require": {
-        "php": ">=5.3.2",
+        "php": ">=5.4",
         "behat/transliterator": "~1.0",
         "doctrine/common": "~2.4"
     },
